chore(storage): remove Go 1.9 code
We no longer support Go 1.9 so the code removed here is dead
and build tags are no longer required.
